[feat]: apply path change
